backend/requirements.txt
Flask
Werkzeug==2.0.1
numpy
requests==2.24.0
gunicorn==20.1.0
scikit-learn
Pillow
tensorflow-cpu
flask-cors
opencv-python